LPS-46899 Extract DDMForm Validator to its own module
